Implement EventTarget interface for WebSocket (Android)
